Added updater_get_file() and updater_fwrite().
updater_get_file() can be used  to get a single file from an FTP server.
It delays the actual destination file creation until the first write
callback so that it won't create an empty file in case the remote file
doesn't exist or something else fails.
updater_fwrite() is the first write callback.