I always wanted thermal reliefs on SMD pads.
Maybe you find it useful.
If so I will adapt it a little that it is more cleaner and
incorporate it in the existing polygon functions (most part
of what I added is copy/pasted).
I don't think a lot of fundamental changes are required, I
only added an extra type in const.h to be able to find the
pads and added some extra arguments to the pad-polygon subtract
function.

Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>